Move inference checker to its own file
